refactor: add subcommands for presentation
